SA1649 File name should match first type name
